Fix bug where SOC fails to decrease at low currents

+++ Core/Inc/main.h
@@ -70,7 +70,7 @@ typedef struct batteryModule {
 	uint16_t hvsens_pack_voltage;
 	uint16_t read_auxreg[NUM_AUXES];
 	uint16_t balance_status[NUM_DEVICES];
-    uint16_t soc;
+    uint32_t soc; // microamps!!!!!
     uint32_t current;
     uint16_t dew_point[NUM_DEVICES];
 } batteryModule;
